+++ clients/llm_client.py
@@ -1,0 +1,18 @@
+from abc import ABC, abstractmethod
+from typing import Optional, Dict, Any
+import base64
+
+class LLMClient(ABC):
+    def encode_image(self, image_path: str) -> str:
+        with open(image_path, "rb") as image_file:
+            return base64.b64encode(image_file.read()).decode('utf-8')
+
+    @abstractmethod
+    def generate(self,
+        prompt: str,
+        image_path: Optional[str] = None,
+        stream: bool = False,
+        model: str = "llama3.2-vision",
+        temperature: float = 0.2,
+        num_predict: int = 256) -> Dict[Any, Any]:
+        pass
