+++ config.py
@@ -1,0 +1,87 @@
+import argparse
+from pathlib import Path
+import json
+from typing import Any
+import logging
+
+logger = logging.getLogger(__name__)
+
+class Config:
+    def __init__(self, config_dir: str = "config"):
+        self.config_dir = Path(config_dir)
+        self.user_config = self.config_dir / "config.json"
+        self.default_config = self.config_dir / "default_config.json"
+        self.load_config()
+
+    def load_config(self):
+        """Load configuration from JSON file with cascade:
+        1. Try user config (config.json)
+        2. Fall back to default config (default_config.json)
+        """
+        try:
+            if self.user_config.exists():
+                logger.info(f"Loading user config from {self.user_config}")
+                with open(self.user_config) as f:
+                    self.config = json.load(f)
+            else:
+                logger.info(f"No user config found, loading default config from {self.default_config}")
+                with open(self.default_config) as f:
+                    self.config = json.load(f)
+        except Exception as e:
+            logger.error(f"Error loading config: {e}")
+            raise
+
+    def get(self, key: str, default: Any = None) -> Any:
+        """Get configuration value with optional default."""
+        return self.config.get(key, default)
+
+    def update_from_args(self, args: argparse.Namespace):
+        """Update configuration with command line arguments."""
+        for key, value in vars(args).items():
+            if value is not None:  # Only update if argument was provided
+                if key == "client":
+                    self.config["clients"]["default"] = value
+                elif key == "ollama_url":
+                    self.config["clients"]["ollama"]["url"] = value
+                elif key == "openrouter_key":
+                    self.config["clients"]["openrouter"]["api_key"] = value
+                    # If key is provided but no client specified, use OpenRouter
+                    if not args.client:
+                        self.config["clients"]["default"] = "openrouter"
+                elif key == "model":
+                    client = self.config["clients"]["default"]
+                    self.config["clients"][client]["model"] = value
+                else:
+                    self.config[key] = value
+
+    def save_user_config(self):
+        """Save current configuration to user config file."""
+        try:
+            self.config_dir.mkdir(parents=True, exist_ok=True)
+            with open(self.user_config, 'w') as f:
+                json.dump(self.config, f, indent=2)
+            logger.info(f"Saved user config to {self.user_config}")
+        except Exception as e:
+            logger.error(f"Error saving user config: {e}")
+            raise
+
+def get_client(config: Config) -> str:
+    """Get the appropriate client based on configuration."""
+    client_type = config.get("clients", {}).get("default", "ollama")
+    client_config = config.get("clients", {}).get(client_type, {})
+    
+    if client_type == "ollama":
+        return client_config.get("url", "http://localhost:11434")
+    elif client_type == "openrouter":
+        api_key = client_config.get("api_key")
+        if not api_key:
+            raise ValueError("OpenRouter API key is required when using OpenRouter client")
+        return api_key
+    else:
+        raise ValueError(f"Unknown client type: {client_type}")
+
+def get_model(config: Config) -> str:
+    """Get the appropriate model based on client type and configuration."""
+    client_type = config.get("clients", {}).get("default", "ollama")
+    client_config = config.get("clients", {}).get(client_type, {})
+    return client_config.get("model", "llama3.2-vision")
