+++ config/default_config.json
@@ -1,24 +1,35 @@
 {
-    "ollama_url": "http://localhost:11434",
-    "model": "llama3.2-vision",
+    "clients": {
+        "default": "ollama",
+        "ollama": {
+            "url": "http://localhost:11434",
+            "model": "llama3.2-vision"
+        },
+        "openrouter": {
+            "api_key": "",
+            "model": "llama3.2-vision"
+        }
+    },
     "prompt_dir": "prompts",
     "output_dir": "output",
-    "frames_per_minute": 60,
-    "whisper_model": "medium",
-    "keep_frames": false,
-    "frame_analysis_threshold": 10.0,
-    "min_frame_difference": 5.0,
-    "max_frames": 30,
+    "frames": {
+        "per_minute": 60,
+        "analysis_threshold": 10.0,
+        "min_difference": 5.0,
+        "max_count": 30
+    },
     "response_length": {
         "frame": 300,
         "reconstruction": 1000,
         "narrative": 500
     },
     "audio": {
+        "whisper_model": "medium",
         "sample_rate": 16000,
         "channels": 1,
         "quality_threshold": 0.2,
         "chunk_length": 30,
         "language_confidence_threshold": 0.8
-    }
+    },
+    "keep_frames": false
 }
